Moved host networking code into a new `host::network` module
