+++ src/main/build.rs
@@ -55,7 +55,7 @@ fn run_cbindgen(build_common: &ShadowBuildCommon) {
                 "main/host/descriptor/descriptor_types.h".into(),
                 "main/host/descriptor/tcp.h".into(),
                 "main/host/futex_table.h".into(),
-                "main/host/network_interface.h".into(),
+                "main/host/network/network_interface.h".into(),
                 "main/host/protocol.h".into(),
                 "main/host/status_listener.h".into(),
                 "main/host/syscall_handler.h".into(),
@@ -369,8 +369,8 @@ fn build_shadow_c(build_common: &ShadowBuildCommon) {
         "host/syscall/unistd.c",
         "host/syscall/uio.c",
         "host/syscall_condition.c",
-        "host/network_interface.c",
-        "host/network_queuing_disciplines.c",
+        "host/network/network_interface.c",
+        "host/network/network_queuing_disciplines.c",
         "host/tracker.c",
         "routing/payload.c",
         "routing/packet.c",
