+++ src/main/host/descriptor/socket/inet/legacy_tcp.rs
@@ -16,10 +16,10 @@ use crate::host::descriptor::{
 };
 use crate::host::host::Host;
 use crate::host::memory_manager::MemoryManager;
+use crate::host::network::net_namespace::NetworkNamespace;
 use crate::host::syscall::io::{write_partial, IoVec};
 use crate::host::syscall_types::{ForeignArrayPtr, SyscallError};
 use crate::host::thread::ThreadId;
-use crate::network::net_namespace::NetworkNamespace;
 use crate::network::packet::PacketRc;
 use crate::utility::callback_queue::{CallbackQueue, Handle};
 use crate::utility::pod;
