+++ src/main/host/descriptor/socket/inet/mod.rs
@@ -10,9 +10,9 @@ use crate::cshadow as c;
 use crate::host::descriptor::socket::{RecvmsgArgs, RecvmsgReturn, SendmsgArgs};
 use crate::host::descriptor::{FileMode, FileState, FileStatus, OpenFile, SyscallResult};
 use crate::host::memory_manager::MemoryManager;
+use crate::host::network::net_namespace::NetworkNamespace;
 use crate::host::syscall::io::IoVec;
 use crate::host::syscall_types::SyscallError;
-use crate::network::net_namespace::NetworkNamespace;
 use crate::network::packet::PacketRc;
 use crate::utility::callback_queue::CallbackQueue;
 use crate::utility::sockaddr::SockaddrStorage;
