+++ src/main/host/host.rs
@@ -35,10 +35,10 @@ use crate::core::work::task::TaskRef;
 use crate::core::worker::Worker;
 use crate::cshadow;
 use crate::host::descriptor::socket::abstract_unix_ns::AbstractUnixNamespace;
-use crate::host::network_interface::{NetworkInterface, PcapOptions};
+use crate::host::network::net_namespace::NetworkNamespace;
+use crate::host::network::network_interface::{NetworkInterface, PcapOptions};
 use crate::host::process::Process;
 use crate::host::thread::ThreadId;
-use crate::network::net_namespace::NetworkNamespace;
 use crate::network::relay::{RateLimit, Relay};
 use crate::network::router::Router;
 use crate::network::PacketDevice;
