+++ src/main/host/network/network_interface.c
@@ -16,8 +16,8 @@
 #include "main/host/descriptor/descriptor.h"
 #include "main/host/descriptor/socket.h"
 #include "main/host/descriptor/tcp.h"
-#include "main/host/network_interface.h"
-#include "main/host/network_queuing_disciplines.h"
+#include "main/host/network/network_interface.h"
+#include "main/host/network/network_queuing_disciplines.h"
 #include "main/host/protocol.h"
 #include "main/host/tracker.h"
 #include "main/routing/address.h"
