+++ src/main/network/mod.rs
@@ -3,7 +3,6 @@ use std::net::Ipv4Addr;
 use crate::network::packet::PacketRc;
 
 pub mod graph;
-pub mod net_namespace;
 pub mod packet;
 pub mod relay;
 pub mod router;
